improved parsing !

+++ src/main/java/estamosremoto/utils/bytecode/BytecodeModel.java
@@ -2,8 +2,8 @@ package estamosremoto.utils.bytecode;
 
 public record BytecodeModel(
         String magic,
-        short minorVersion,
-        short majorVersion,
-        short constantPoolCount
+        int minorVersion,
+        int majorVersion,
+        int constantPoolCount
 ) {
 }
